+++ qom/loopers/XYZLooper.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.loopers.XYZLooper'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-12-28'
-__updated__ = '2021-08-20'
+__updated__ = '2021-08-23'
 
 # dependencies
 from typing import Union
@@ -28,16 +28,17 @@ class XYZLooper(BaseLooper):
         Function to loop, formatted as ``func(system_params, val, logger, results)``, where ``system_params`` is a dictionary of the parameters for the system, ``val`` is the current value of the looping parameter, ``logger`` is an instance of the module logger and ``results`` is a list of tuples each containing ``val`` and the value calculated within the function.
     params : dict
         Parameters for the looper and optionally, the system and the plotter. The "looper" key is a dictionary containing the keys "X", "Y" and "Z", each with the keys "var" and "val" for the name of the parameter to loop and its corresponding values, along with additional options (refer notes).
-
+    cb_update : callable, optional
+        Callback function to update status and progress, formatted as ``cb_update(status, progress, reset)``, where ``status`` is a string, ``progress`` is an integer and ``reset`` is a boolean.
 
     .. note:: All the options defined under the "looper" dictionary in ``params`` supersede individual function arguments. Refer :class:`qom.loopers.BaseLooper` for a complete list of supported options.
     """
 
-    def __init__(self, func, params: dict, cb_progress=None):
+    def __init__(self, func, params: dict, cb_update=None):
         """Class constructor for XYZLooper."""
 
         # initialize super class
-        super().__init__(func=func, params=params, code='xyz_looper', name='3D Looper', cb_progress=cb_progress)
+        super().__init__(func=func, params=params, code='xyz_looper', name='3D Looper', cb_update=cb_update)
 
         # set axes
         self._set_axis(axis='X')
@@ -46,6 +47,10 @@ class XYZLooper(BaseLooper):
 
         # display initialization
         logger.info('-------------------Looper Initialized-----------------\t\n')
+
+        # update callback
+        if self.cb_update is not None:
+            self.cb_update(status='Looper Initialized', progress=None)
 
     def loop(self, grad: bool=False, grad_position='all', grad_mono_idx: int=0, mode: str='serial'):
         """Method to calculate the output of a given function for each X-axis Y-axis and Z-axis point.
@@ -171,5 +176,9 @@ class XYZLooper(BaseLooper):
 
         # display completion
         logger.info('-------------------Results Obtained-------------------\t\n')
+
+        # update callback
+        if self.cb_update is not None:
+            self.cb_update(status='Results Obtained', progress=None, reset=True)
 
         return self.results