+++ qom/ui/widgets/LooperWidget.py
@@ -6,11 +6,10 @@
 __name__    = 'qom.ui.widgets.LooperWidget'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-08-19'
-__updated__ = '2021-08-20'
+__updated__ = '2021-08-23'
 
 # dependencies
 from PyQt5 import QtCore, QtGui, QtWidgets
-import importlib
 import logging
 import os
 
@@ -27,8 +26,14 @@ class LooperWidget(BaseWidget):
     
     Parameters
     ----------
-    parent : QtWidget.*
+    parent : :class:`qom.ui.GUI`
         Parent class for the widget.
+    solver_widget : :class:`qom.ui.widgets.SolverWidget`
+        Solver widget.
+    system_widget : :class:`qom.ui.widgets.SystemWidget`
+        System widget.
+    plotter_widget : :class:`qom.ui.widgets.PlotterWidget`
+        Plotter widget.
     """
 
     def __init__(self, parent, solver_widget, system_widget, plotter_widget):
@@ -53,13 +58,14 @@ class LooperWidget(BaseWidget):
         
         # frequently used variables
         width = 640
+        height = 240
         row_height = 32
         padding = 32
 
         # fix size
         self.setFixedWidth(width)
-        # move under header
-        self.move(640, 240 + 4)
+        # set as central widget on the right with margin
+        self.move(width, height + 4)
 
         # initialize UI elements
         # looper label
@@ -69,55 +75,45 @@ class LooperWidget(BaseWidget):
         # function combo box
         self.cmbx_func = QtWidgets.QComboBox()
         self.cmbx_func.setFixedSize(width / 2 - 1.5 * padding, row_height)
-        self.cmbx_func.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
-        self.cmbx_func.currentTextChanged.connect(self.on_cmbx_func_text_changed)
+        self.cmbx_func.currentTextChanged.connect(self.set_curr_func)
         self.cmbx_func.setVisible(False)
         # wrap check box
         self.chbx_wrap = QtWidgets.QCheckBox('Wrap')
         self.chbx_wrap.setFixedSize(width / 4, row_height)
-        self.chbx_wrap.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
         self.chbx_wrap.setVisible(False)
         # plot check box
         self.chbx_plot = QtWidgets.QCheckBox('Plot')
         self.chbx_plot.setFixedSize(width / 4, row_height)
-        self.chbx_plot.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
         self.chbx_plot.setVisible(False)
         # file path line edit
         self.le_path = QtWidgets.QLineEdit('')
         self.le_path.setFixedSize(width / 2 - 1.5 * padding, row_height)
-        self.le_path.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
         self.le_path.setVisible(False)
         # mode combo box
         self.cmbx_mode = QtWidgets.QComboBox()
         self.cmbx_mode.setFixedSize(width / 4 - 1.25 * padding, row_height)
-        self.cmbx_mode.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
         self.cmbx_mode.addItems(['serial', 'multithread'])
         self.cmbx_mode.setVisible(False)
         # loop button
         self.btn_loop = QtWidgets.QPushButton('Loop')
         self.btn_loop.setFixedSize(width / 4 - 1.25 * padding, row_height)
-        self.btn_loop.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
         self.btn_loop.clicked.connect(self.loop)
         self.btn_loop.setVisible(False)
         # variable label
         self.lbl_var = QtWidgets.QLabel('var')
         self.lbl_var.setFixedSize(width / 4 - 1.25 * padding, row_height)
-        self.lbl_var.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
         self.lbl_var.setVisible(False)
         # minimum value label
         self.lbl_min = QtWidgets.QLabel('min')
         self.lbl_min.setFixedSize(width / 4 - 1.25 * padding, row_height)
-        self.lbl_min.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
         self.lbl_min.setVisible(False)
         # maximum value label
         self.lbl_max = QtWidgets.QLabel('max')
         self.lbl_max.setFixedSize(width / 4 - 1.25 * padding, row_height)
-        self.lbl_max.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
         self.lbl_max.setVisible(False)
         # dimension label
         self.lbl_dim = QtWidgets.QLabel('dim')
         self.lbl_dim.setFixedSize(width / 4 - 1.25 * padding, row_height)
-        self.lbl_dim.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
         self.lbl_dim.setVisible(False)
 
         # update layout 
@@ -161,9 +157,15 @@ class LooperWidget(BaseWidget):
         # disable loop button
         self.btn_loop.setDisabled(True)
 
+        # handle no function found
+        if self.cmbx_func.currentText() == 'NA':
+            # update status
+            self.parent.update(status='No function found')
+            return
+        
         # get function
         func = get_looper_func(self.system_widget.system, {}, self.cmbx_func.currentText())
-        # extract looper parameters
+        # extract axes parameters from widgets
         looper_params = dict()
         axes = ['X', 'Y', 'Z']
         keys = ['var', 'min', 'max', 'dim']
@@ -172,7 +174,7 @@ class LooperWidget(BaseWidget):
             for j in range(4):
                 looper_params[axes[i]][keys[j]] = self.param_widgets[i * 4 + j].text()
 
-        # get file path
+        # get file path to save/load data
         path_text = self.le_path.text()
         file_path = path_text if path_text != '' else None
 
@@ -182,35 +184,21 @@ class LooperWidget(BaseWidget):
             'solver': self.solver_widget.get_params(),
             'system': self.system_widget.get_params(),
             'plotter': self.plotter_widget.get_params()
-        }, cb_progress=self.parent.update_progress)
-
-        # update status
-        self.parent.update_status('Looping...')
-
-        # # run looper on different thread
-        # import threading
-        # _thread = threading.Thread(target=looper.wrap, args=(file_path, True, ))
-        # _thread.start()
-        # _thread.join()
+        }, cb_update=self.parent.update)
 
         # run looper on main thread
         looper.wrap(file_path=file_path, plot=True)
 
-        # update status
-        self.parent.update_status('Result Obtained')
-        # update progress
-        self.parent.reset_progress()
-
         # enable loop button
         self.btn_loop.setDisabled(False)
 
-    def on_cmbx_func_text_changed(self, value):
-        """Method to update the UI when combo box selection changes.
+    def set_curr_func(self, value):
+        """Method to update the widget when combo box selection changes.
         
         Parameters
         ----------
         value : str
-            New text of the combo box
+            New text of the combo box.
         """
 
         # frequently used variables
@@ -221,17 +209,25 @@ class LooperWidget(BaseWidget):
         # enable button
         self.btn_loop.setDisabled(False)
 
-        # file path
-        self.le_path.setText('data/' + self.system_widget.system({}).code + '/' + value)
+        # handle no function found
+        if self.cmbx_func.currentText() == 'NA':
+            # update status
+            self.parent.update(status='No function found')
+            return
 
-        # get looper type
-        dim = self.loopers.index(self.looper) + 1
-        found = lambda s, d: sum([0 if s.find(e) == -1 else 1 for e in expr[:d + 1]]) == d + 1 and sum([0 if s.find(e) == -1 else 1 for e in expr]) == d + 1
+        # set data file path
+        system_code = self.system_widget.system({}).code if self.system_widget.system is not None else 'sys'
+        self.le_path.setText('data/' + system_code + '/' + value)
 
+        # update parameters
         params = {}
-        # check for scripts directory
+        # search script name by matching looper dimension
+        found = lambda name, dim: sum([0 if name.find(e) == -1 else 1 for e in expr[:dim + 1]]) == dim + 1 and sum([0 if name.find(e) == -1 else 1 for e in expr]) == dim + 1
+        # search scripts
         for script_name in os.listdir('scripts'):
-            if found(script_name, dim):
+            # if script found
+            if found(script_name, self.loopers.index(self.looper) + 1):
+                # read lines
                 with open('scripts/' + script_name) as script_file:
                     lines = script_file.readlines()
                     line_start = 0
@@ -243,24 +239,26 @@ class LooperWidget(BaseWidget):
                         param_lines += lines[i]
                         if lines[i].find('}') == 0:
                             break
+                    # evaluate to dictionary
                     params = eval(param_lines[9:])
-        
+                    # close file
+                    script_file.close()
+                    
+        # update looper parameters
         if 'looper' in params:
             for i in range(int(len(self.param_widgets) / 4)):
                 for j in range(4):
                     if keys[j] in params['looper'][axes[i]]:
                         self.param_widgets[i * 4 + j].setText(str(params['looper'][axes[i]][keys[j]])) 
-        
+        # update system parameters
         if 'system' in params:
-            self.system_widget.set_params(params['system'])
-        
+            self.system_widget.set_params(params=params['system'])
+        # update plotter parameters
         if 'plotter' in params:
-            self.plotter_widget.set_params(params['plotter'])
+            self.plotter_widget.set_params(params=params['plotter'])
 
-        # update status
-        self.parent.update_status('Ready')
-        # update progress
-        self.parent.reset_progress()
+        # update footer
+        self.parent.update(status='Ready', progress=None, reset=True)
 
     def set_curr_item(self, pos):
         """Method to set the current looper.
@@ -285,45 +283,25 @@ class LooperWidget(BaseWidget):
         for widget in self.param_widgets:
             self.layout.removeWidget(widget)
             widget.deleteLater()
+            
+        # reset list
         self.param_widgets = list()
 
         # get combo box items
-        cmbx_items = [self.system_widget.cmbx_func.itemText(i) for i in range(self.system_widget.cmbx_func.count())] if self.system_widget.cmbx_func.itemText(0) != 'NA' else []
+        cmbx_items = ['NA'] if self.system_widget.cmbx_func.count() == 0 or self.system_widget.cmbx_func.currentText() == 'NA' else [self.system_widget.cmbx_func.itemText(i) for i in range(self.system_widget.cmbx_func.count())]
+
+        # # get wrapped functions
         # default_looper_func_names_dict = importlib.import_module('qom.utils.looper', '*').__dict__.get('default_looper_func_names_dict')
         # cmbx_items += [key for key in default_looper_func_names_dict]
 
         # update layout
         widget_col = pre_count * 4
-        for i in range(pos + 1):
-            # var widget
-            le_var = QtWidgets.QLineEdit('')
-            le_var.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
-            le_var.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
-            le_var.setFixedSize(width / 4 - 1.25 * padding, row_height)
-            # min widget
-            le_min = QtWidgets.QLineEdit('')
-            le_min.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
-            le_min.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
-            le_min.setFixedSize(width / 4 - 1.25 * padding, row_height)
-            # max widget
-            le_max = QtWidgets.QLineEdit('')
-            le_max.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
-            le_max.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
-            le_max.setFixedSize(width / 4 - 1.25 * padding, row_height)
-            # dim widget
-            le_dim = QtWidgets.QLineEdit('')
-            le_dim.setFont(QtGui.QFont('Segoe UI', pointSize=9, italic=False))
-            le_dim.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
-            le_dim.setFixedSize(width / 4 - 1.25 * padding, row_height)
-            # update list
-            self.param_widgets.append(le_var)
-            self.param_widgets.append(le_min)
-            self.param_widgets.append(le_max)
-            self.param_widgets.append(le_dim)
-            self.layout.addWidget(le_var, int(widget_col / 4), int(widget_col % 4) + 0, 1, 1, alignment=QtCore.Qt.AlignRight)
-            self.layout.addWidget(le_min, int(widget_col / 4), int(widget_col % 4) + 1, 1, 1, alignment=QtCore.Qt.AlignRight)
-            self.layout.addWidget(le_max, int(widget_col / 4), int(widget_col % 4) + 2, 1, 1, alignment=QtCore.Qt.AlignRight)
-            self.layout.addWidget(le_dim, int(widget_col / 4), int(widget_col % 4) + 3, 1, 1, alignment=QtCore.Qt.AlignRight)
+        for _ in range(pos + 1):
+            for j in range(4):
+                widget = QtWidgets.QLineEdit('')
+                widget.setFixedSize(width / 4 - 1.25 * padding, row_height)
+                self.param_widgets.append(widget)
+                self.layout.addWidget(widget, int(widget_col / 4), int(widget_col % 4) + j, 1, 1, alignment=QtCore.Qt.AlignRight)
             # update count
             widget_col += 4
 
@@ -349,9 +327,13 @@ class LooperWidget(BaseWidget):
         Parameters
         ----------
         theme : str, optional
-            Display theme:
-                'dark': Dark mode.
-                'light': Light mode.
+            Display theme. Available options are:
+                ==========  ==============
+                value       meaning
+                ==========  ==============  
+                "dark"      dark mode.
+                "light"     light mode.
+                ==========  ==============
         """
 
         # update theme