+++ qom/ui/widgets/PlotterWidget.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.ui.widgets.PlotterWidget'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2021-08-20'
-__updated__ = '2021-08-20'
+__updated__ = '2021-08-23'
 
 # dependencies
 from PyQt5 import QtCore, QtGui, QtWidgets
@@ -24,7 +24,7 @@ class PlotterWidget(BaseWidget):
     
     Parameters
     ----------
-    parent : QtWidget.*
+    parent : :class:`qom.ui.GUI`
         Parent class for the widget.
     """
 
@@ -62,7 +62,6 @@ class PlotterWidget(BaseWidget):
         # plotter parameters
         self.te_params = QtWidgets.QTextEdit('')
         self.te_params.setFixedSize(width - 2 * padding, row_height * 5)
-        self.te_params.setFont(QtGui.QFont('Segoe UI', pointSize=10, italic=False))
 
         # update layout 
         self.layout = QtWidgets.QGridLayout()
@@ -91,7 +90,15 @@ class PlotterWidget(BaseWidget):
         return codes
     
     def get_params(self):
-        # get parameters
+        """Method to obtain the parameters for the plotter.
+        
+        Returns
+        -------
+        params: dict
+            Parameters for the plotter.
+        """
+
+        # evaluate parameters
         params = eval(self.te_params.toPlainText()) if self.te_params.toPlainText() != '' else {}
         
         return params
@@ -113,6 +120,14 @@ class PlotterWidget(BaseWidget):
         self.lbl_name.setText('Plotter Parameters:')
     
     def set_params(self, params):
+        """Method to set the parameters for the plotter.
+        
+        Parameters
+        ----------
+        params: dict
+            Parameters for the plotter.
+        """
+        
         # set parameters
         self.te_params.setText(str(params))
 
@@ -122,9 +137,13 @@ class PlotterWidget(BaseWidget):
         Parameters
         ----------
         theme : str, optional
-            Display theme:
-                'dark': Dark mode.
-                'light': Light mode.
+            Display theme. Available options are:
+                ==========  ==============
+                value       meaning
+                ==========  ==============  
+                "dark"      dark mode.
+                "light"     light mode.
+                ==========  ==============
         """
 
         # update theme